[FEAT] 컴포넌트 분리 | Separate components for better modularity TALEARNT-307
