+++ src/modal/TalentsSettingModal/core/talentsSettingModal.api.ts
@@ -1,6 +1,6 @@
 import { postAPI } from "@utils/apiMethods";
 
-import { talentsType } from "@modal/TalentsSettingModal/talentsSettingModal.type";
+import { talentsType } from "@modal/TalentsSettingModal/core/talentsSettingModal.type";
 
 export const postTalents = async (talents: Record<talentsType, number[]>) =>
   await postAPI("/v1/users/my-talents", talents, { withCredentials: true });
