+++ src/modal/TalentsSettingModal/core/talentsSettingModal.type.ts
@@ -1,0 +1,6 @@
+export type talentsType = "giveTalents" | "receiveTalents";
+
+export type talentsDataType = Record<
+  talentsType,
+  { label: string; value: number }[]
+>;
