--- src/modal/TalentsSettingModal/talentsSettingModal.type.ts
@@ -1,1 +1,0 @@
-export type talentsType = "giveTalents" | "receiveTalents";
